+++ lib-data-stream-redis/src/main/groovy/io/seqera/data/stream/MessageConsumer.groovy
@@ -88,7 +88,7 @@ package io.seqera.data.stream
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  * @since 1.0
  * @see MessageStream#consume(String, MessageConsumer)
- * @see AbstractMessageStream
+ * @see AbstractMessageStreamConsumer
  */
 interface MessageConsumer<T> {
